feat(#1039): Graph storage integration + affinity consolidation + test fixes (#1059)
* fix: Fix mypy type errors in calendar connector

- Update read_content, write_content to return HandlerResponse
- Update delete_content, content_exists to return HandlerResponse
- Update get_content_size, get_ref_count to return HandlerResponse
- Update is_directory, mkdir, rmdir to return HandlerResponse
- Fix ConfirmLevel comparison methods type signatures
- Update tests and E2E script to unwrap HandlerResponse


* fix: Remove unused type: ignore comment for yaml import


* feat(#1039): Add graph storage layer for entities and relationships

Implements PostgreSQL-native graph storage for GraphRAG:
- GraphStore class with entity/relationship CRUD operations
- N-hop neighbor traversal via recursive CTEs
- Subgraph extraction for context building
- Entity resolution via embedding similarity
- Multi-tenant isolation

Changes:
- Add src/nexus/search/graph_store.py with GraphStore implementation
- Add store_to_graph parameter to Memory.store() API
- Add graph API endpoints: /api/graph/entity, /api/graph/search, etc.
- Add alembic migration for entities, relationships, entity_mentions tables
- Add integration and e2e tests


* fix(#1039): Fix CI failures - lint errors and file size exception

- Remove f-string without placeholders in test scripts
- Remove unused imports (json, Session)
- Prefix unused arguments with underscore
- Add memory_api.py to file size exceptions (2097 lines)


* style(#1039): Apply ruff formatting to test scripts


* fix(#1039): Fix Alembic migration dependency to use merged head

Update graph storage migration to depend on 6e9842c71775 (merged head)
instead of add_relationship_extraction to resolve multiple heads error.


* feat(#1026): Add affinity-based memory consolidation and fix integration tests

## Affinity Consolidation (Issue #1026)
- Add SimpleMem-inspired affinity scoring combining semantic similarity and temporal proximity
- Formula: affinity = β · cos(v_i, v_j) + (1 - β) · exp(-λ|t_i - t_j|)
- Add agglomerative clustering for memory grouping
- Add consolidate_by_affinity_async() method to ConsolidationEngine
- Add numpy and scikit-learn dependencies

## Test Fixes
- Fix skills lifecycle tests to match implementation behavior (tier param is legacy/ignored)
- Fix tiger resource map sync tests (skip tests requiring PostgreSQL on SQLite)
- Fix e2e tests to use FastAPI mode for Graph API support
- Fix Graph API asyncio.run() bug - use await instead in async endpoints

## Other Fixes
- Fix timestamp timezone consistency in metadata_store.py (ensure UTC)
- Fix CompactFileMetadata to use float timestamps (preserve microseconds)


* fix(#1026): Fix lint errors in test_affinity.py

- Use datetime.UTC instead of timezone.utc (UP017)
- Remove intermediate assignment that caused E402


* fix(#1026): Fix remaining lint errors

- Move _ensure_utc function after imports in metadata_store.py (E402)
- Fix import sorting in gmail_oauth.py (I001)
- Apply ruff formatting to gmail_connector.py and test_consolidation.py


---------

+++ src/nexus/core/compact_metadata.py
@@ -188,8 +188,8 @@ class CompactFileMetadata:
     size: int
     hash_id: int  # -1 for None
     mime_id: int  # -1 for None
-    created_ts: int  # Unix timestamp, 0 for None
-    modified_ts: int  # Unix timestamp, 0 for None
+    created_ts: float  # Unix timestamp with microseconds, 0.0 for None
+    modified_ts: float  # Unix timestamp with microseconds, 0.0 for None
     version: int
     tenant_id: int  # -1 for None
     user_id: int  # -1 for None
@@ -214,9 +214,9 @@ class CompactFileMetadata:
         tenant_id = _tenant_pool.intern(metadata.tenant_id) if metadata.tenant_id else -1
         user_id = _user_pool.intern(metadata.created_by) if metadata.created_by else -1
 
-        # Convert datetimes to Unix timestamps
-        created_ts = int(metadata.created_at.timestamp()) if metadata.created_at else 0
-        modified_ts = int(metadata.modified_at.timestamp()) if metadata.modified_at else 0
+        # Convert datetimes to Unix timestamps (preserve microseconds as float)
+        created_ts = metadata.created_at.timestamp() if metadata.created_at else 0.0
+        modified_ts = metadata.modified_at.timestamp() if metadata.modified_at else 0.0
 
         # Pack boolean flags
         flags = 0
